requirements.txt
attrs==20.3.0
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
google-api-core==1.16.0
google-auth==1.11.2
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
idna==2.9
iniconfig==1.1.1
packaging==20.9
pathtools==0.1.2
pluggy==0.13.1
powerline-status==2.7
prompt-toolkit==3.0.2
protobuf==3.11.3
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
Pygments==2.5.2
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
pytz==2019.3
quocofs-python
requests==2.23.0
rsa==4.0
six==1.14.0
toml==0.10.2
urllib3==1.25.8
watchdog==0.10.4
wcwidth==0.1.8
